add error boundary to the dashboard refactor widgets and dashboard components
