+++ src/pages/dashboards/grid/widget-area.js
@@ -19,7 +19,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import Dropzone from "./dropzone";
 import Grid from "./grid";
-import WidgetFactory from "../widget/widget-factory.jsx";
+import WidgetFactory from "../widget/widget-factory";
 
 class WidgetArea extends React.Component {
   snapToGrid(value) {
