+++ src/store/apiSlice.js
@@ -115,5 +115,4 @@ export const {
   useDeleteUserFromUsergroupMutation,
   useUpdateUsergroupMutation,
   useGetWidgetsQuery,
-  useLazyGetICbyIdQuery,
 } = apiSlice;
